♻️ Move HTTP handlers from main.go to handlers.go
